requirements.txt
Flask==3.0.2
yt-dlp==2024.3.10
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
Flask-Caching==2.1.0
gunicorn==21.2.0